argtype for icons are globalised
