+++ src/lib/components/BusinessCard/BusinessCard.stories.tsx
@@ -1,11 +1,9 @@
 import BusinessCard from "./BusinessCard";
 import { Meta, StoryObj } from "@storybook/nextjs";
-import { iconOptions, iconDecorator } from "../../../utils/storybookUtils.tsx";
 
 const meta: Meta<typeof BusinessCard> = {
   title: "Components/BusinessCard",
   component: BusinessCard,
-  decorators: [iconDecorator],
   argTypes: {
     position: { table: { defaultValue: { summary: "center" } } },
     variant: { table: { defaultValue: { summary: "secondary" } } },
@@ -25,11 +23,6 @@ const meta: Meta<typeof BusinessCard> = {
           onClick: () => alert("Button clicked!"),
         },
       },
-    },
-    icon: {
-      options: Object.keys(iconOptions),
-      mapping: iconOptions,
-      control: { type: "select" },
     },
   },
   args: {
