+++ src/lib/components/Button/Button.stories.tsx
@@ -1,21 +1,14 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
 
 import Button from "./Button";
-import { iconOptions, iconDecorator } from "../../../utils/storybookUtils.tsx";
 
 const meta: Meta<typeof Button> = {
   title: "Components/Button",
   component: Button,
-  decorators: [iconDecorator],
   argTypes: {
     variant: { table: { defaultValue: { summary: "primary" } } },
     size: { table: { defaultValue: { summary: "md" } } },
     shape: { table: { defaultValue: { summary: "solid" } } },
-    icon: {
-      options: Object.keys(iconOptions),
-      mapping: iconOptions,
-      control: { type: "select" },
-    },
   },
   args: {
     label: "Button",
