+++ src/lib/components/Chip/Chip.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
 
 import Chip from "./Chip";
-import { iconOptions, iconDecorator } from "../../../utils/storybookUtils.tsx";
 
 const meta: Meta<typeof Chip> = {
   title: "Components/Chip",
@@ -11,18 +10,12 @@ const meta: Meta<typeof Chip> = {
     shape: { table: { defaultValue: { summary: "solid" } } },
     size: { table: { defaultValue: { summary: "md" } } },
     pill: { table: { defaultValue: { summary: "true" } } },
-    icon: {
-      options: Object.keys(iconOptions),
-      mapping: iconOptions,
-      control: { type: "select" },
-    },
   },
   args: {
     label: "Chips",
     pill: true,
     closable: true,
   },
-  decorators: [iconDecorator],
 };
 
 export default meta;
