+++ src/lib/components/Dropdown/Dropdown.stories.tsx
@@ -1,21 +1,14 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
 import Dropdown from "./Dropdown";
-import { iconOptions, iconDecorator } from "../../../utils/storybookUtils.tsx";
 
 const meta: Meta<typeof Dropdown> = {
   title: "Components/Dropdown",
   component: Dropdown,
-  decorators: [iconDecorator],
   argTypes: {
     shape: { table: { defaultValue: { summary: "solid" } } },
     variant: { table: { defaultValue: { summary: "primary" } } },
     size: { table: { defaultValue: { summary: "md" } } },
     spacing: { table: { defaultValue: { summary: "callout" } } },
-    icon: {
-      options: Object.keys(iconOptions),
-      mapping: iconOptions,
-      control: { type: "select" },
-    },
   },
   args: {
     label: "Dropdown Menu",
