+++ src/lib/components/ListView/Item/ListViewItem.stories.tsx
@@ -1,23 +1,9 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
 import ListView from "../ListView";
-import { iconOptions, iconDecorator } from "../../../../utils/storybookUtils.tsx";
 
 const meta: Meta<typeof ListView.Item> = {
   title: "Components/ListView/ListView.Item",
   component: ListView.Item,
-  decorators: [iconDecorator],
-  argTypes: {
-    icon: {
-      options: Object.keys(iconOptions),
-      mapping: iconOptions,
-      control: { type: "select" },
-    },
-    iconRight: {
-      options: Object.keys(iconOptions),
-      mapping: iconOptions,
-      control: { type: "select" },
-    },
-  },
   args: {
     title: "My Document Folder",
     description: "Personal notes and other files",
